norm, ifndef, paco --s [handle -1], delete a.out test.c, function name

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: nkiticha <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/03 17:43:23 by nkiticha          #+#    #+#             */
-/*   Updated: 2023/11/23 16:27:27 by nkiticha         ###   ########.fr       */
+/*   Updated: 2023/11/26 17:15:20 by nkiticha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 
 int	ft_printf_c(char c);
 int	ft_printf_s(char *str);
-int	ft_printf_p(unsigned long long n);
+int	ft_printf_p(unsigned long n);
 int	ft_printf_di(int n);
 int	ft_printf_u(unsigned int n);
 int	ft_printf_x(unsigned int n);
